Uploaded program's .py file and .txt doc
Initial commit/upload for a project that I had previously "completed" and will likely update later. Commit includes an uploaded python file that holds the code of the program itself as well as a text document meant to be used as the original plan/outline. The text document includes a sample run and the pseudocode that the program was built from.